requirements.txt
Django~=3.1.2
djangorestframework~=3.14.0
django-colorfield ~= 0.8.0
psycopg2-binary~=2.9.6